chore: add test for about.py to avoid failed codecov checks when releasing
 - if about.py changes and we don't test it, codecov is all like "wah, you didn't hit your diff targets because you went from 0% to 0% on about.py" :sweat_smile:
